Test.Kore.Builtin.AssocComm.CeilSimplifier: Factor out
